Delete exteraneous files and move them to the UnityRouter class
